@@ -36,7 +36,7 @@
         "@sentry/webpack-plugin": "^1.10.0",
         "@testing-library/dom": "^7.1.2",
         "@testing-library/jest-dom": "^5.14.1",
-        "@testing-library/react": "^10.0.1",
+        "@testing-library/react": "^14.0.0",
         "@testing-library/user-event": "^14.4.3",
         "@types/file-saver": "^2.0.2",
         "@types/history": "^4.7.8",
@@ -47,8 +47,8 @@
         "@types/node": "^14.17.3",
         "@types/number-to-words": "^1.2.1",
         "@types/prop-types": "^15.7.3",
-        "@types/react": "^17.0",
-        "@types/react-dom": "^17.0",
+        "@types/react": "^18.2.24",
+        "@types/react-dom": "^18.2.8",
         "@types/react-dropzone": "^4.2.2",
         "@types/react-router": "^5.1.15",
         "@types/react-router-dom": "^5.1.7",
@@ -168,8 +168,8 @@
         "moment-timezone": "^0.5.16",
         "number-to-words": "^1.2.4",
         "prop-types": "^15.7.2",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
         "react-dropzone": "^6.2.4",
         "react-helmet-async": "^1.2.3",
         "react-image": "^4.1.0",
