@@ -1,9 +1,9 @@
 import React from 'react';
-import { fireEvent, render, WithReduxStore } from 'test-utils';
+import { fireEvent, render, WithReduxStore, waitFor, waitForElementToBeRemoved } from 'test-utils';
 import Immutable from 'immutable';
 import { AddToFavouritesButton } from './AddToFavouritesButton';
 import * as redux from 'react-redux';
-import { act } from 'react-dom/test-utils';
+import { act } from '@testing-library/react';
 
 const setup = ({ state = {}, props = {} } = {}) => {
     return render(
@@ -33,9 +33,13 @@ describe('AddToFavouritesButton', () => {
         const { queryByTestId, getByTestId } = setup({
             props: { selectedJournals: { 1: true }, clearSelectedJournals: jest.fn() },
         });
-        await act(async () => await fireEvent.click(getByTestId('add-to-favourites-button')));
+
+        fireEvent.click(getByTestId('add-to-favourites-button'));
+        await waitFor(() => getByTestId('add-to-favourites-confirmation-box'));
         expect(getByTestId('add-to-favourites-confirmation-box')).toBeInTheDocument();
-        await act(async () => await jest.advanceTimersByTime(2000));
-        expect(queryByTestId('add-to-favourites-confirmation-box')).not.toBeInTheDocument();
+        await act(() => jest.advanceTimersByTime(2000));
+        await waitForElementToBeRemoved(getByTestId('add-to-favourites-confirmation-box')).then(() =>
+            expect(queryByTestId('add-to-favourites-confirmation-box')).not.toBeInTheDocument(),
+        );
     });
 });
